requirements.txt
altair==5.5.0
Bio==1.8.0
biopython==1.85
gseapy==1.1.9
h5py==3.14.0
joblib==1.4.2
matplotlib==3.10.3
networkx==3.4.2
numpy==2.3.1
opencv_python==4.11.0.86
pandas==2.3.1
Pillow==11.3.0
plotly==6.2.0
python_louvain==0.16
Requests==2.32.4
scikit_learn==1.7.0
scipy==1.16.0
seaborn==0.13.2
statsmodels==0.14.5
streamlit==1.46.1
tensorflow==2.19.0
torch==2.5.1
torchvision==0.20.1
